pyproject.toml
[project]
name = "simple-employee-manager"
version = "0.1.0"
description = "Warehouse Employee Management System - Desktop application for tracking employees, certifications, and compliance"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "peewee>=3.17.0",
    "python-dateutil>=2.8.0",
    "openpyxl>=3.1.0",
    "typer[all]>=0.12.0",
    "rich>=13.7.0",
    "tabulate>=0.9.0",
    "questionary>=2.0.0",
    "customtkinter>=5.2.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]
build = [
    "pyinstaller>=6.0.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "freezegun>=1.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
pythonpath = "src"
